pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Educational project using Flask, PostgreSQL, and deployment on a PaaS service"
authors = ["Egor Efimov <odes36@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
Flask-WTF = "^1.2.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
